License correction
`LICENSE` file indicates BSD 3-Clause, but this file (`setup.py`) previously listed the project is MIT licensed. Both licenses are great, but mixing the two may cause confusion, including in packaged versions of this software. This proposed change presumes that you intended the project to be BSD-licensed, and corrects the setup.py license classifer and  license field.